webscraper/requirements.txt
requests
newspaper3k
ddgs
lxml_html_clean
urllib3<2
